Create CompanyCard component, clean up code
